ContentDialog cleanup
