+++ FluentAvalonia/UI/Controls/ContentDialog/ContentDialogButton.cs
@@ -1,10 +1,28 @@
 ﻿namespace FluentAvalonia.UI.Controls
 {
+    /// <summary>
+    /// Defines constants that specify the default button on a content dialog.
+    /// </summary>
     public enum ContentDialogButton
     {
+        /// <summary>
+        /// No button is specified as the default.
+        /// </summary>
         None = 0,
+
+        /// <summary>
+        /// The primary button is the default.
+        /// </summary>
         Primary = 1,
+
+        /// <summary>
+        /// The secondary button is the default.
+        /// </summary>
         Secondary = 2,
+
+        /// <summary>
+        /// The close button is the default.
+        /// </summary>
         Close = 3
     }
 }
