--- FluentAvalonia/UI/Controls/ContentDialog/ContentDialogClosedEventArgs.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace FluentAvalonia.UI.Controls
-{
-    public class ContentDialogClosedEventArgs : EventArgs
-    {
-        public ContentDialogResult Result { get; }
-
-        public ContentDialogClosedEventArgs(ContentDialogResult res)
-        {
-            Result = res;
-        }
-    }
-}
