+++ FluentAvalonia/UI/Controls/ContentDialog/ContentDialogClosingDeferral.cs
@@ -1,9 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace FluentAvalonia.UI.Controls
+﻿namespace FluentAvalonia.UI.Controls
 {
+    /// <summary>
+    /// Represents a deferral that can be used by an app to respond asyncronously to
+    /// a <see cref="ContentDialog"/> closing
+    /// </summary>
     public class ContentDialogClosingDeferral
     {
         internal ContentDialogClosingDeferral(ContentDialog owner)
@@ -11,6 +11,9 @@ namespace FluentAvalonia.UI.Controls
             _owner = owner;
         }
 
+        /// <summary>
+        /// Notifies the system that the app has finished processing the closing event.
+        /// </summary>
         public void Complete()
         {
             _owner.CompleteClosingDeferral();
