+++ FluentAvalonia/UI/Controls/ContentDialog/EventArgs/ContentDialogButtonClickEventArgs.cs
@@ -2,6 +2,9 @@
 
 namespace FluentAvalonia.UI.Controls
 {
+    /// <summary>
+    /// Provides data for the button click events.
+    /// </summary>
     public class ContentDialogButtonClickEventArgs : EventArgs
     {
         internal ContentDialogButtonClickEventArgs(ContentDialog owner)
@@ -9,12 +12,16 @@ namespace FluentAvalonia.UI.Controls
             _owner = owner;
         }
 
+        /// <summary>
+        /// Gets or sets a value that can cancel the button click. 
+        /// A true value for Cancel cancels the default behavior.
+        /// </summary>
         public bool Cancel { get; set; }
 
         internal bool IsDeferred => _deferral != null;
 
         /// <summary>
-        /// Gets a ContentDialogButtonClickDeferral the app can use to respond
+        /// Gets a <see cref="ContentDialogButtonClickDeferral"/> the app can use to respond
         /// asyncronously to a button click event
         /// </summary>
         public ContentDialogButtonClickDeferral GetDeferral()
