+++ FluentAvalonia/UI/Controls/ContentDialog/EventArgs/ContentDialogClosingEventArgs.cs
@@ -2,6 +2,9 @@
 
 namespace FluentAvalonia.UI.Controls
 {
+    /// <summary>
+    /// Provides data for the closing event.
+    /// </summary>
     public class ContentDialogClosingEventArgs : EventArgs
     {
         internal ContentDialogClosingEventArgs(ContentDialog owner, ContentDialogResult res)
@@ -10,12 +13,24 @@ namespace FluentAvalonia.UI.Controls
             _owner = owner;
         }
 
+        /// <summary>
+        /// Gets or sets a value that can cancel the closing of the dialog.
+        /// A true value for Cancel cancels the default behavior.
+        /// </summary>
         public bool Cancel { get; set; }
 
+        /// <summary>
+        /// Gets the <see cref="ContentDialogResult"/> of the closing event.
+        /// </summary>
         public ContentDialogResult Result { get; }
 
         internal bool IsDeferred => _deferral != null;
 
+        /// <summary>
+        /// Gets a <see cref="ContentDialogClosingDeferral"/> that the app can use to 
+        /// respond asynchronously to the closing event.
+        /// </summary>
+        /// <returns></returns>
         public ContentDialogClosingDeferral GetDeferral()
         {
             _deferral = new ContentDialogClosingDeferral(_owner);
